rename BillingSyncConfig
